Namespace concern module names
The modules defined in app/models/concerns are all references as if they
are namespaced when defined (they aren't)

e.g.

Defined as

    module SkipRequireCardNumbersConcern
      ...
    end

referenced as

    Spree::CreditCard.include SolidusBraintree::SkipRequireCardNumbersConcern

This commit namespaces the concerns when they are defined.

I'm not too sure how this previously worked, however in an upgrade to
Solidus 1.1.1 my store couldn't properly resolve the constant name any
more.