Change solidus dependency to components
The gem currently depends on solidus (which is a meta gem for all
solidus modules). This means that even if the site does not use
solidus_frontend, this gem will require it to be installed.

This changes the dependency to only include solidus_core and solidus_api
which are the components this actually requires.