Add config to disable ship address verification
Braintree supports address verification if an address is present,
however, this can sometimes lead to addresses being declined because
they don't match the address on file with the PayPal account or credit
card.

As some people may wish to disable this, I have added this configuration
to be able to disable sending the adderss to Braintree.