Update how braintree client tokens are generated
To promote a one to one relationship between braintree customers and
spree users, add the braintree customer id to the spree_users table.
Anytime a new client token is generated for a new payment method, pass
that customer id in as one of the options so that all new payment
methods are added to the same braintree customer associated to the spree
user.