Replace OrderWalkthrough with local version
The extension currently modifies the address factory so that it can
set a fixed zipcode for the addresses created by OrderWalkthrough.
However, once we change the frontend of this gem to
SolidusStarterFrontend and run this extension's specs together with the
frontend specs, the address modification would cause the frontend's
coupon code spec to fail. See
https://app.circleci.com/pipelines/github/solidusio/solidus_braintree/239/workflows/06370bde-ef25-49cf-8fc2-1aefd5a2f86e/jobs/623/parallel-runs/0/steps/0-108.

To prevent that, we're going to avoid modifying the address by having a
local version of the OrderWalkthrough class.