Fix load order issue with `table_name_prefix`
When running `bin/rails g solidus:install` and selecting `braintree` as
the payment method, there is currently a failure during the running of
the `db:seeds` task that points to an issue with the table prefix not
taking effect

```
ActiveRecord::StatementInvalid: Could not find table 'solidus_braintree_configurations' (ActiveRecord::StatementInvalid)
```

which comes from the store decorator here `app/decorators/models/solidus_braintree/spree_store_decorator.rb:14:in `build_default_configuration'`.

After some extensive testing, it seems like this is a load order issue
in development (autoloading enabled) where the definition needs to
happen before the engine is declared. The previous location of this
method happens after the `engine.rb` require.

The cleanest solution here seems to be to declare the table name in each
model as there are only 3 currently in the extension, so we've opted for
that.

Co-authored-by: Benjamin Willems <benjamin@super.gd>